Exit immediately on raising error
The _run error handling in the _run function was causing it
to continue onto gather results even after TestRun.run had
failed, but was trying to use a variable that hadn't been created
due to the error handing. This commit simply changes _run to
exit on an error instead of continuing on to resul parsing.