Add built-in mode config file for Survey tool (#758)
* add changes to fresh branch

* add survey mode config

* add minimial changes back in

* add missing types

* remove trailing whitespace

* add hello_world c program to test survey

* Fixed hanks problems.

* Fix bug when error.request is None

* Fix error in hello_c.yaml

* Partially fix Opt and Sopt plugins

Opt and Sopt plugins are necessary for generating command line
options generated from mode config files. Both Opt and Sopt take
an arguments which can be either a string or a list of strings,
which complicates their arg specs. This commit:

- Adds test cases for both Opt and Sopt plugins.
- Introduces a UnionSpec class to support arguments that are
  a union of multiple types.
- Begins to integrate UnionSpec into argument validation code.

Note that argument validation has not been fully implemented for
UnionSpec, which means that the added tests are not yet passing.

* Finish fixing Opt and Sopt function plugins

Opt and Sopt plugins now have correct behavior for lists of strings
and pass all associated tests. Because both plugins can take either
a single string or a list of strings, this commit also finishes
implementing the UnionSpec specification (introduced in the previous
commit), which allows for function plugins to have parameters whose
actual arguments can take on one of several types. This specification
object is designed in the same spirit of the recently introduced Opt
specification, and will hopefully be useful for future plugins.

This commit also modifies the argument validation logic for function
plugins to account for the UnionSpec specification and adds unit tests
as well.

* Implement spec description for UnionSpec

* Add function plugin for flag

This commit adds a corresponding function plugin for the flag function,
implemented previously. The flag function is used to generate UNIX-style
option flags from boolean-like config options. At present, the flag
function accepts a string as its first argument, rather than a boolean,
the justification being that some versions of YAML specify multiple
conventions for boolean values, including true/false, yes/no, and on/off.
This design decision may need to be revisited at a later date.

* Fix syntax error

* Fix syntax error and mismatched file names

* Add tests for Flag plugin

* Fix bug in UnionSpec __contains__ method

* Replace UnionSpec argspec with MaybeList

Per Paul, UnionSpec gives users too much power. Instead,
replaced it with MaybeList, which can take either a type
or a list of that same type.

* Fix unit test for sopt plugin

* Progress towards JSON output

* More progress towards mode config

* Correct list access in mode config

* Add more unit tests for survey mode config

* Fix bug in JSON parser

* Make all parsers return log

* Pass all results tests

* Make unit tests wait for test run to finish

* Pass style tests

* Implement review feedback

* Attempt to pass CI tests

* Remove all survey-related functionality

Per Paul Ferrell, LANL will no longer provide Survey as a product.
This commit strips out all Survey-related functionality, but
retains several features and bug fixes implemented during development
of the feature. These include:

- A builtin directory, containing a modes subdirectory, presently empty,
which is checked by when resolving tests, and in which future builtin
modes can be placed.

- Expression functions for building command line arguments and options
from lists and Booleans.

- A fix to a bug in the JSON parser which causes it to incorrectly
return an empty dictionary under certain circumstances.

- Other miscellaneous changes and fixes.

* Add gitignore to force inclusion of empty directory

---------

Co-authored-by: Paul Ferrell <51765748+Paul-Ferrell@users.noreply.github.com>
Co-authored-by: Paul Ferrell <pferrell@lanl.gov>