Prevent tests from transitioning to BUILD_CREATED state after failure (#813)
* Prevent tests from transition to BUILD_CREATED state after failure

Previously, tests were failing during the build phase and correctly
logging a `BUILD_ERROR` state, but then incorrectly continuing on to
log a `BUILD_COMPLETE` state, sometimes multiple times. This commit
fixes that. It also fixes a double logging of an ABORTED state (albeit
with different messages).

Several issues with logging states remain:

- In the aforementioned double ABORTED state, one log statement should
probably be removed.

- Nearly identical tests will show either an ENV_FAILED state, or an
ENV_FAILED state followed by a BUILD_FAILED state. This inconsistency
needs to be resolved, but a decision needs to be made as to the
intended behavior.

- Some log states appear to be used inappropriately, which should
eventually be fixed.

* Fix failing unit tests